requirements.txt
django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
djoser==2.3.1
geopy==2.4.1
drf-nested-routers==0.94.2
pillow==11.2.1
djangorestframework_simplejwt==5.5.0
drf_spectacular==0.28.0